Contact Mapper

+++ src/main/java/com/shofydrop/repository/ContactRepository.java
@@ -3,5 +3,6 @@ package com.shofydrop.repository;
 import com.shofydrop.entity.Contact;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContactRepository  extends JpaRepository<Contact, Integer> {
+public interface ContactRepository  extends JpaRepository<Contact, Long> {
+
 }
